Makefile: Drop dependency on locally installed RPM packages
When compiling for the "aarch64" target architecture, the Makefile was
designed to pick up header and library references from the local
directories containing unpacked aarch64 RPM packages.

This gets in the way of building natively on an aarch64 target (such
as the Raspberry Pi and the RockPro64) and also makes the build
process harder to migrate to other distributions.

As a step towards removing the dependency on the locally unpacked
packages, update the Makefile to use standard locations for the header
and library files as reported by the compiler in use - native or
cross.

With the changes in this commit, it is possible to build stage 1 using -

    make stage1

artifacts without the need for the locally unpacked RPM
packages.

TODO:
* Improve "/usr/include" and "/usr/include/$(arch)-linux-gnu"
* Test generated artifacts

Signed-off-by: Punit Agrawal <pagrawal@lynx.com>