conf/aarch64: Use inline atomics
Recent versions of the compiler gained the ability to use outline
atomics to allow runtime detection of LSE on the target
system. Although a useful feature, it doesn't work on systems where
the standard library does not have support for detecting this feature.

Let's consistently decide to use inline atomics. The downside is that
we cannot take advantage of LSE atomics if the system supports it but
it is a reasonable choice for portablity.

Signed-off-by: Punit Agrawal <pagrawal@lynx.com>